lib - clean up suspense/usePromise impl
