+++ packages/lib/src/hooks/index.ts
@@ -6,6 +6,7 @@ export * from "./useEffectEvent.js"
 export * from "./useId.js"
 export * from "./useLayoutEffect.js"
 export * from "./useMemo.js"
+export * from "./usePromise.js"
 export * from "./useReducer.js"
 export * from "./useRef.js"
 export * from "./useState.js"
